+++ CHANGELOG.rst
@@ -4,6 +4,14 @@ CHANGELOG
 Development
 -----------
 
+- [BREAKING, Remove] Healpix related columns from schema and code
+  (`MR105 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/105>`__)
+- [Remove] Unused dependencies from pyproject.toml
+  (`MR105 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/105>`__)
+
+- [Deprecate] HTML upload page is deprecated
+  (`MR105 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/105>`__)
+
 - [Update] Improve error reporting during CSV file validation
   (`MR108 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/108>`__)
 - [Fix] Allow only a single metadata file to be uploaded